comments added, you need to refine comments in functions.cpp

+++ functions.h
@@ -17,13 +17,10 @@
 using namespace std;
 
 
-string send(string);
-string fun(string*, string*, string*, string*);
-void ADC_UART_enabling();
-bool FileExists(const char*);
-float T_ADC_reading();
-
-
-
+string send(string);									// Function for sending of commands to GSM module
+string fun(string*, string*, string*, string*);			// Function for splitting of SMS on variables
+void ADC_UART_enabling();								// Function for enabling of ADC and UART port
+bool FileExists(const char*);							// Function for checking on existence of files
+double T_ADC_reading();									// Function for reading of ADC value
 
 #endif /* FUNCTIONS_H_ */
